♻️refactor: 포맷팅 적용 및 파일 이동
